+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Changelog 
+# Changelog
 
 ## Unreleased
 
 - Add wgpu triangle example ([#53] by [@lord])
+- Add thread-safe app handle ([#90] by [@clavin])